requirements.txt
pytest~=7.4.3
selenium~=4.15.0
Faker~=21.0.0
opencv-python
requests~=2.31.0
ChromeController~=0.3.26
allure-pytest